requirements-vercel.txt
langchain-core>=0.3.0
langchain-groq>=0.3.0
langchain>=0.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
pydantic>=2.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
typing-extensions>=4.5.0
requests>=2.28.0
starlette>=0.27.0
httpx>=0.24.0
